- Har endret alt med tall til BIGDECIMAL, det var veldig rotete med så mange forskjellige datatyper. (Long på id er fortsatt der) - Har gjort ferdig OrderController og OrderService - Laget Input og Output dto´s slik at det ble mer ryddig. - MÅ gjøre små endringer i customer slik at vi bruker dto når vi finner by id og findall. - VI MÅ FIKSE PÅ DB_MIGRATION SAMMEN fordi JEG har endret datatypene i entities, MEN HAR IKKE RØRT I dbmigration enda. REDD for å fucke up.

+++ src/main/java/internal/dtos/ProductDto.java
@@ -2,11 +2,13 @@ package internal.dtos;
 
 import internal.enums.ProductStatus;
 
+import java.math.BigDecimal;
+
 public record ProductDto(
         String name,
         String description,
-        Double price,
+        BigDecimal price,
         ProductStatus productStatus,
-        Long quantity
+        BigDecimal quantity
 ) {
 }
